+++ config/individual.js
@@ -54,7 +54,6 @@ const components = [
 ];
 const plugins = [
   'alert',
-  'bus',
   'confirm',
   'event',
   'grid',